requirements.txt
# Python 3.13.7 compatible (LangChain w/ Pydantic v2)
numpy>=2.0
pydantic>=2.8
langchain>=0.2.20
langchain-core>=0.3.0
langchain-nvidia-ai-endpoints>=0.2.1
python-dotenv>=1.0.1
kagglehub>=0.3.0